+++ src/main/java/es/ubu/lsi/ubumonitor/clustering/controller/collector/ActivityCollector.java
@@ -26,7 +26,7 @@ public class ActivityCollector extends DataCollector {
 			for (CourseModule courseModule : selected) {
 				ActivityCompletion activity = courseModule.getActivitiesCompletion().get(userData.getEnrolledUser());
 				State state = activity.getState();
-				double value = (state == State.COMPLETE || state == State.COMPLETE_PASS) ? 1.0 : 0.0;
+				int value = (state == State.COMPLETE || state == State.COMPLETE_PASS) ? 1 : 0;
 				userData.addDatum(new Datum(getType(), courseModule.getModuleName(),
 						courseModule.getModuleType().getModName(), value));
 				userData.addNormalizedDatum(value);
